+++ src/test/scala/org/constellation/regression/DoubleSpendValidationTest.scala
@@ -1,7 +1,6 @@
 package org.constellation.regression
 
 import java.security.KeyPair
-
 import cats.effect.IO
 import cats.effect.concurrent.Ref
 import cats.syntax.all._
@@ -20,6 +19,8 @@ import org.mockito.{ArgumentMatchersSugar, IdiomaticMockito}
 import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.should.Matchers
 
+import scala.concurrent.ExecutionContext
+
 // TODO: Consider moving to validation test suite
 class DoubleSpendValidationTest
     extends AnyFreeSpec
@@ -28,7 +29,10 @@ class DoubleSpendValidationTest
     with IdiomaticMockitoCats
     with ArgumentMatchersSugar {
 
-  def go()(implicit dao: DAO): GenesisObservation = Genesis.createGenesisObservation(Seq.empty)
+  implicit val concurrent = IO.ioConcurrentEffect(IO.contextShift(ExecutionContext.global))
+  implicit val cs = IO.contextShift(ExecutionContext.global)
+
+  def go(dao: DAO): GenesisObservation = Genesis.createGenesisObservation(Seq.empty)(dao).unsafeRunSync()
 
   def startingTips(go: GenesisObservation)(implicit dao: DAO): Seq[SignedObservationEdge] =
     Seq(go.initialDistribution.soe, go.initialDistribution2.soe)
@@ -89,7 +93,7 @@ class DoubleSpendValidationTest
         as.lookup(*) shouldReturnF AddressCacheData(200L, 200L, balanceByLatestSnapshot = 200L).some
         dao.transactionService.isAccepted(alreadyAcceptedTx.hash) shouldReturnF true
 
-        val cb = CheckpointBlock.createCheckpointBlockSOE(txs, startingTips(go()))
+        val cb = CheckpointBlock.createCheckpointBlockSOE(txs, startingTips(go(dao)))
 
         val isValidCb = cbv.simpleValidation(cb).map(_.isValid)
 
@@ -106,7 +110,7 @@ class DoubleSpendValidationTest
         as.lookup(*) shouldReturnF None
         dao.transactionService.isAccepted(*) shouldReturnF false
 
-        val cb = CheckpointBlock.createCheckpointBlockSOE(txs, startingTips(go()))
+        val cb = CheckpointBlock.createCheckpointBlockSOE(txs, startingTips(go(dao)))
 
         val isValidCb = cbv.simpleValidation(cb).map(_.isValid)
         isValidCb.unsafeRunSync shouldBe true