+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Changed
+- Genesis acceptance rewritten to depend on generic effect type
+
+## [Unreleased]
+### Changed
 - Better tx destination address validation (parity number and base58 characters)
 - Update whitelisting file